requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
supabase>=1.0.3
python-dotenv>=1.0.0
requests>=2.28.2
pydantic>=1.10.7
